Feat/Refactor (Renderer): Deletion queue and rendering on another image.
Frames are now being rendered onto another image before being copied to the swap chain.
A deletion queue has been implemented.
Removed the VulkanWrapper folder and put everything back in VulkanRenderer.
