+++ FlashlightEngine/Include/FlashlightEngine/VulkanRenderer/VulkanTypes.hpp
@@ -17,6 +17,7 @@
 
 #include <glm/mat4x4.hpp>
 #include <glm/vec4.hpp>
+#include <ranges>
 
 namespace Flashlight {
     enum class DebugLevel {
@@ -28,6 +29,31 @@ namespace Flashlight {
         Debug = 4
     };
 
+    struct DeletionQueue {
+        std::deque<std::function<void()>> Deletors;
+
+        void PushFunction(std::function<void()>&& function) {
+            Deletors.push_back(std::move(function));
+        }
+
+        void Flush() {
+            // Reverse iterate the deletion queue to execute all the functions.
+            for (auto& deletor : std::ranges::reverse_view(Deletors)) {
+                deletor();
+            }
+
+            Deletors.clear();
+        }
+    };
+
+    struct AllocatedImage {
+        VkImage Image;
+        VkImageView ImageView;
+        VmaAllocation Allocation;
+        VkExtent3D ImageExtent;
+        VkFormat ImageFormat;
+    };
+
     struct FrameData {
         VkCommandPool CommandPool;
         VkCommandBuffer MainCommandBuffer;
@@ -36,6 +62,8 @@ namespace Flashlight {
         VkFence RenderFence;
 
         u32 SwapchainImageIndex;
+
+        DeletionQueue DeletionQueue;
     };
 }
