+++ FlashlightEngine/Source/VulkanRenderer/VulkanUtils/VulkanImageUtils.cpp
@@ -11,7 +11,8 @@
 #include <FlashlightEngine/VulkanRenderer/VulkanInitializers.hpp>
 
 namespace Flashlight::VulkanRenderer::VulkanUtils {
-    void TransitionImage(const VkCommandBuffer commandBuffer, const VkImage image, const VkImageLayout currentLayout,
+    void TransitionImage(const VkCommandBuffer commandBuffer, const VkImage image,
+                         const VkImageLayout currentLayout,
                          const VkImageLayout newLayout) {
         VkImageMemoryBarrier2 imageBarrier = {.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2};
         imageBarrier.pNext = nullptr;
@@ -40,4 +41,39 @@ namespace Flashlight::VulkanRenderer::VulkanUtils {
 
         vkCmdPipelineBarrier2(commandBuffer, &depInfo);
     }
+
+    void CopyImageToImage(const VkCommandBuffer commandBuffer, const VkImage source, const VkImage destination,
+                          const VkExtent2D srcSize, const VkExtent2D dstSize) {
+        VkImageBlit2 blitRegion{.sType = VK_STRUCTURE_TYPE_IMAGE_BLIT_2, .pNext = nullptr};
+
+        blitRegion.srcOffsets[1].x = static_cast<i32>(srcSize.width);
+        blitRegion.srcOffsets[1].y = static_cast<i32>(srcSize.height);
+        blitRegion.srcOffsets[1].z = 1;
+
+        blitRegion.dstOffsets[1].x = static_cast<i32>(dstSize.width);
+        blitRegion.dstOffsets[1].y = static_cast<i32>(dstSize.height);
+        blitRegion.dstOffsets[1].z = 1;
+
+        blitRegion.srcSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
+        blitRegion.srcSubresource.baseArrayLayer = 0;
+        blitRegion.srcSubresource.layerCount = 1;
+        blitRegion.srcSubresource.mipLevel = 0;
+
+        blitRegion.dstSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
+        blitRegion.dstSubresource.baseArrayLayer = 0;
+        blitRegion.dstSubresource.layerCount = 1;
+        blitRegion.dstSubresource.mipLevel = 0;
+
+        VkBlitImageInfo2 blitInfo{.sType = VK_STRUCTURE_TYPE_BLIT_IMAGE_INFO_2, .pNext = nullptr};
+        blitInfo.dstImage = destination;
+        blitInfo.dstImageLayout = VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL;
+        blitInfo.srcImage = source;
+        blitInfo.srcImageLayout = VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL;
+        blitInfo.filter = VK_FILTER_LINEAR;
+        blitInfo.regionCount = 1;
+        blitInfo.pRegions = &blitRegion;
+
+        vkCmdBlitImage2(commandBuffer, &blitInfo);
+    }
+
 }
